% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evCM.R
\name{evCM}
\alias{evCM}
\title{Confusion Matrix Calculation Function}
\usage{
evCM(Ypred, Ytrue, nclasses, CM.print = T, sample.name = "")
}
\arguments{
\item{Ypred}{a vector of predicted values.}

\item{Ytrue}{a vector of actual values.}

\item{nclasses}{the number of actual classes.}

\item{CM.print}{a logical argument; whether to print the confusion matrix on the console or not.}

\item{sample.name}{is an optional character string specifying the name of the sample.}
}
\description{
This function calculates a confusion matrix for the provided predicted (in rows) and true (in columns) values. It is also obligatory to provide the actual number of classes. If some of the classes are missed, the function adds the necessary number of zero rows (in case of predictions) or zero columns (in case of true classes) to the classification matrix.
}
\examples{
evCM(Ypred, Ytrue)
}
\keyword{(confusion)}
\keyword{classification}
\keyword{matrix}
