% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evCMetrics.R
\name{evCMetrics}
\alias{evCMetrics}
\title{Function for CM metrics calculation}
\usage{
evCMetrics(CM, cat.names = vector(mode = "character", length = 0),
  s.expname = "")
}
\arguments{
\item{CM}{a confusion matrix.}

\item{cat.names}{a vector of category names (optional).}

\item{s.expname}{character string giving the name of experiment (optional).}
}
\description{
This function calculates a number of specified metrics (precision, recall, and f1-score) for a given confusion matrix. R^2 count is also calculated.
}
\details{
Important! The confusion matrix should have no missing rows or columns and should be square!
}
\examples{
evCMetrics(myCM)
}
\keyword{classification}
\keyword{metrics}
