% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/evStudy.R
\name{evStudy}
\alias{evStudy}
\title{An Event Study Analysis Function}
\usage{
evStudy(df.events, df.stocks, event.window = 2, estim.window = 100)
}
\arguments{
\item{df.events}{The data frame containing the list of events.}
}
\description{
This function runs event study calculations using the standard market model.
}
\examples{
evStudy()
}
\keyword{analysis}
\keyword{event}
\keyword{study}

